@@ -13,7 +13,7 @@
         }
     ],
     "require": {
-        "php": ">=7.4.0",
+        "php": "^8.1",
         "ext-pdo_sqlite": "*",
         "guzzlehttp/guzzle": "^7.5",
         "slim/slim": "^3.12",
@@ -23,6 +23,7 @@
         "matthiasmullie/minify": "^1.3",
         "ork/phpcs": "^2.8",
         "phpstan/phpstan": "^1.8",
+        "rector/rector": "^0.14.6",
         "roave/security-advisories": "dev-latest"
     },
     "autoload": {
@@ -37,7 +38,8 @@
             "minifycss css/bgg.css > public/bgg.min.css"
         ],
         "phpcs": "phpcs",
-        "phpstan": "phpstan"
+        "phpstan": "phpstan",
+        "rector": "rector --clear-cache --dry-run"
     },
     "config": {
         "allow-plugins": {
